refactor: 모임 기도 불러오는 기능 GroupPrayController로 이동
