+++ src/main/java/com/uspray/uspray/controller/PrayController.java
@@ -4,7 +4,6 @@ package com.uspray.uspray.controller;
 import com.uspray.uspray.DTO.ApiResponseDto;
 import com.uspray.uspray.DTO.pray.PrayListResponseDto;
 import com.uspray.uspray.DTO.pray.request.PrayRequestDto;
-import com.uspray.uspray.DTO.pray.request.PrayToGroupPrayDto;
 import com.uspray.uspray.DTO.pray.request.PrayUpdateRequestDto;
 import com.uspray.uspray.DTO.pray.response.PrayResponseDto;
 import com.uspray.uspray.exception.SuccessStatus;
@@ -153,19 +152,5 @@ public class PrayController {
     ) {
         return ApiResponseDto.success(SuccessStatus.CANCEL_PRAY_SUCCESS,
             prayFacade.cancelPray(prayId, user.getUsername()));
-    }
-
-    @Operation(summary = "모임 기도제목으로 불러오기")
-    @ApiResponse(
-        responseCode = "200",
-        description = "모임 기도제목으로 불러오기",
-        content = @Content(schema = @Schema(implementation = PrayResponseDto.class)))
-    @PostMapping("/pray-to-grouppray")
-    public ApiResponseDto<?> prayToGroupPray(
-        @RequestBody PrayToGroupPrayDto prayToGroupPrayDto,
-        @Parameter(hidden = true) @AuthenticationPrincipal User user
-    ) {
-        prayFacade.prayToGroupPray(prayToGroupPrayDto, user.getUsername());
-        return ApiResponseDto.success(SuccessStatus.PRAY_TO_GROUP_PRAY_SUCCESS);
     }
 }
